+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sum incomes/expenditures from different adservers in billing history
 ### Fixed
 - Grouping events by day in billing history
+- Importing banners when external classifier is not defined
 
 ## [1.2.5] - 2019-10-01
 ### Added